% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOP_models_benchmark.R
\name{L23_IOP}
\alias{L23_IOP}
\alias{R18_IOP}
\title{Benchmark IOP models}
\usage{
L23_IOP(
  aph,
  wavelen = wavelen_IOP,
  rng1 = runif(1),
  rng2 = runif(1),
  rng3 = runif(1),
  rng4 = runif(1),
  rng5 = runif(1),
  rng6 = runif(1),
  rng7 = runif(1),
  rng8 = runif(1)
)

R18_IOP(Chl, ag440, TSM, wavelen = wavelen_IOP)
}
\arguments{
\item{aph}{Absorption coefficient of phytoplankton in the unit of [m^-1]}

\item{wavelen}{Wavelength in the unit of [nm] with the same length of \code{aph}}

\item{rng1}{Random value between 0 and 1}

\item{rng2}{Random value between 0 and 1}

\item{rng3}{Random value between 0 and 1}

\item{rng4}{Random value between 0 and 1}

\item{rng5}{Random value between 0 and 1}

\item{rng6}{Random value between 0 and 1}

\item{rng7}{Random value between 0 and 1}

\item{rng8}{Random value between 0 and 1}

\item{Chl}{Chlorophyll a concentration, as a proxy of biomass concentration
of phytoplankton, in the unit [mg/m^3]}

\item{ag440}{CDOM absorption coefficient at 440 nm, as a proxy of
concentration of CDOM (or called gelbstoff), in the unit [m^-1]}

\item{TSM}{Total suspended matter concentration in the unit [g/m3]}
}
\value{
A list including spectral of IOPs and key parameters
}
\description{
Benchmark IOP models
}
