% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOP_models_main.R
\name{IOP_four_comp}
\alias{IOP_four_comp}
\alias{IOP_two_comp}
\title{Four-term and Two-term IOP models by Bi et al. (202x)}
\usage{
IOP_four_comp(
  Chl = 1,
  ag440 = 0.18,
  ISM = 1,
  Temp = 20,
  Sal = 15,
  qt_bd = 0.5,
  qt_md = 0.5,
  frac_phyto = rand_frac(),
  S_cdom = 0.017422,
  bbdtilde = 0.0216,
  wavelen = wavelen_IOP,
  A_d = NULL,
  G_d = NULL,
  lib_cdom = TRUE,
  aw_smooth = FALSE,
  aw_version = 3,
  ag_seed = NULL,
  ...
)

IOP_two_comp(
  Chl,
  frac_phyto = NULL,
  Temp = 20,
  Sal = 15,
  wavelen = wavelen_IOP,
  aw_smooth = FALSE,
  aw_version = 3,
  seed = NULL,
  ...
)
}
\arguments{
\item{Chl}{Chlorophyll a concentration, as a proxy of biomass concentration
of phytoplankton, in the unit [mg/m^3]}

\item{ag440}{CDOM absorption coefficient at 440 nm, as a proxy of
concentration of CDOM (or called gelbstoff), in the unit [m^-1]}

\item{ISM}{Inorganic suspended matter concentration in the unit [g/m3]}

\item{Temp}{Water temperature in the unit of [degC]}

\item{Sal}{Water salinity in the unit of [PSU]}

\item{qt_bd}{Quantile for biogenic detritus absorption, default as 0.5}

\item{qt_md}{Quantile for minerogenic detritus absorption, default as 0.5}

\item{frac_phyto}{Fraction of phytoplankton groups. Vector with a length of 7.}

\item{S_cdom}{Spectral slope for CDOM absorption coefficient. Not used if
\code{lib_cdom} is \code{TRUE}}

\item{bbdtilde}{Backscattering probability of detritus}

\item{wavelen}{Wavelength in the unit of [nm]}

\item{A_d}{Scaled albedo parameters of detritus}

\item{G_d}{Power law exponent of spectral attenuation of detritus}

\item{lib_cdom}{Option to use the library of CDOM absorption. Default as
\code{TRUE}}

\item{aw_smooth}{Option to smooth the pure water absorption. Default as
\code{FALSE}}

\item{aw_version}{Pure water absorption version: 1, 2, and 3}

\item{ag_seed}{Random seed for ag model. Default as 1234. Fixed to reproduce
the same outputs.}

\item{...}{Parameters passed to \code{IOP_ph_B22_C2}}

\item{seed}{Seed number of the two-term model}
}
\value{
A list including spectral of IOPs and key parameters
}
\description{
Four-term and Two-term IOP models by Bi et al. (202x)
}
\examples{

str(IOP_four_comp(Chl = 1, ag440 = 0.18, ISM = 1))
str(IOP_two_comp(Chl = 1))
}
