% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOPmodel-package.R
\docType{package}
\name{IOPmodel-package}
\alias{IOPmodel}
\alias{IOPmodel-package}
\title{IOPmodel: Model inherent optical properties from component concentrations}
\description{
Supplyment code package for our paper Bi et al. (2023) "Bio-geo-optical modeling of natural waters" submitted to "Fontiers in Marine Science". The package includes our proposed IOP model and also other benchmark models.
}
\author{
\strong{Maintainer}: Shun Bi \email{Shun.Bi@hereon.de} (\href{https://orcid.org/0000-0002-2668-5808}{ORCID})

}
\keyword{internal}
