% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WOPP.R
\name{WOPP}
\alias{WOPP}
\title{WOPP}
\usage{
WOPP(Temp = 25, Sal = 10, wavelen = seq(350, 900, 5))
}
\arguments{
\item{Temp}{Temperature [degC]}

\item{Sal}{Salinity [PSU]}

\item{wavelen}{Wavelength [nm]}
}
\value{
A list of IOPs of pure water
}
\description{
WOPP
}
\note{
The WOPP code is wrapper from Schoenfeld's codes

This code is copied from https://github.com/bishun945/RrsTrans/blob/master/R/WOPP.R
}
\examples{
WOPP(Temp = 25, Sal = 10)
}
\references{
Röttgers, R, R Doerffer, D McKee, and W Schönfeld. “The Water Optical
Properties Processor (WOPP): Pure Water Spectral Absorption, Scattering and
Real Part of Refractive Index Model.” Technical Report No WOPP-ATBD/WRD6,
2016. https://calvalportal.ceos.org/tools.
}
