% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOP_models_benchmark.R
\name{L23_IOP}
\alias{L23_IOP}
\alias{R18_IOP}
\title{Benchmark IOP models}
\usage{
L23_IOP(
  aph,
  wavelen = wavelen_IOP,
  rng1 = runif(1),
  rng2 = runif(1),
  rng3 = runif(1),
  rng4 = runif(1),
  rng5 = runif(1),
  rng6 = runif(1),
  rng7 = runif(1),
  rng8 = runif(1),
  L23update = TRUE
)

R18_IOP(Chl, ag440, TSM, wavelen = wavelen_IOP)
}
\arguments{
\item{aph}{Absorption coefficient of phytoplankton in the unit of [m^-1]}

\item{wavelen}{Wavelength in the unit of [nm] with the same length of \code{aph}}

\item{rng1}{Random value between 0 and 1}

\item{rng2}{Random value between 0 and 1}

\item{rng3}{Random value between 0 and 1}

\item{rng4}{Random value between 0 and 1}

\item{rng5}{Random value between 0 and 1}

\item{rng6}{Random value between 0 and 1}

\item{rng7}{Random value between 0 and 1}

\item{rng8}{Random value between 0 and 1}

\item{L23update}{The option for updated L23 model setups.
According to https://doi.org/10.5194/essd-2023-80-AC1, the p2 parameter
is revised to allow higher values (up to 0.9 rather than 0.5), which allows
higher NAP proportion from aph440. Set L23update as TRUE will have higher
anap in the simulation result. Note that the parameter will be eventually
deleted once the model is in the final version.}

\item{Chl}{Chlorophyll a concentration, as a proxy of biomass concentration
of phytoplankton, in the unit [mg/m^3]}

\item{ag440}{CDOM absorption coefficient at 440 nm, as a proxy of
concentration of CDOM (or called gelbstoff), in the unit [m^-1]}

\item{TSM}{Total suspended matter concentration in the unit [g/m3]}
}
\value{
A list including spectral of IOPs and key parameters
}
\description{
Benchmark IOP models
}
\references{
Loisel H, Jorge D S F, Reynolds R A, et al. A synthetic database generated by
radiative transfer simulations in support of studies in ocean optics and
optical remote sensing of the global ocean\link{J}. Earth System Science Data
Discussions, 2023: 1-24.

Ramírez‐Pérez M, Twardowski M, Trees C, et al. Inversion of in situ light
absorption and attenuation measurements to estimate constituent
concentrations in optically complex shelf seas\link{J}. Journal of Geophysical
Research: Oceans, 2018, 123(1): 720-737.
}
